# [PERF] product: Speed-up `product.template._name_search`