@pivi-odoo 2025-01-13
The `_name_search` was reworked into `_search_display_name` in 18. The issue shouldn't appear on that version, as only the explicit `name_search` is possible to trigger it, but it has a limit but default (and most of the time from the front-end)

@robodoo up to saas-17.4